Fix yield across metamethod/C-call boundary
Since lua-memcached requires that we use the access_by* lua directive,
and since we cannot use multiple access_by* directives in the same
location, we have to merge the two into a single call.

Unfortunately, the straightforward implementation using `dofile` also
isn't supported in the lua/jit version(s) we are using.

The fix is to lift these calls into actual modules and then call the
modules from a single access_by*.

This change also updates tests to take this into account